+++ src/pyspacewar/tests/test_ui.py
@@ -1,4 +1,5 @@
 #!/usr/bin/env python
+from __future__ import print_function
 
 import unittest
 import doctest
@@ -18,8 +19,8 @@ class SurfaceStub(object):
 def doctest_Viewport():
     """Tests for Viewport
 
-        >>> from world import Vector
-        >>> from ui import Viewport
+        >>> from pyspacewar.world import Vector
+        >>> from pyspacewar.ui import Viewport
         >>> viewport = Viewport(SurfaceStub())
         >>> viewport.origin
         Vector(0, 0)
@@ -63,8 +64,8 @@ def doctest_Viewport():
 def doctest_Viewport_screen_size_change():
     """Tests for Viewport
 
-        >>> from world import Vector
-        >>> from ui import Viewport
+        >>> from pyspacewar.world import Vector
+        >>> from pyspacewar.ui import Viewport
         >>> viewport = Viewport(SurfaceStub())
 
     ``screen_pos`` converts world coordinates to screen coordinates
@@ -87,8 +88,8 @@ def doctest_Viewport_screen_size_change():
 def doctest_Viewport_keep_visible():
     """Tests for Viewport.keep_visible.
 
-        >>> from world import Vector
-        >>> from ui import Viewport
+        >>> from pyspacewar.world import Vector
+        >>> from pyspacewar.ui import Viewport
         >>> viewport = Viewport(SurfaceStub())
 
     Points that are already visible change nothing
@@ -96,7 +97,7 @@ def doctest_Viewport_keep_visible():
         >>> viewport.keep_visible([Vector(0, 0)], 10)
         >>> viewport.keep_visible([Vector(100, 100)], 10)
         >>> viewport.keep_visible([Vector(-200, -100)], 10)
-        >>> print viewport.origin, viewport.scale
+        >>> print(viewport.origin, viewport.scale)
         (0.000, 0.000) 1.0
 
     We can see the range of points that are inside the view margin, for
@@ -110,15 +111,15 @@ def doctest_Viewport_keep_visible():
     Points that are off-screen cause scrolling
 
         >>> viewport.keep_visible([Vector(600, 200)], 10)
-        >>> print viewport.origin, viewport.scale
+        >>> print(viewport.origin, viewport.scale)
         (210.000, 0.000) 1.0
 
         >>> viewport.keep_visible([Vector(300, 700)], 10)
-        >>> print viewport.origin, viewport.scale
+        >>> print(viewport.origin, viewport.scale)
         (210.000, 410.000) 1.0
 
         >>> viewport.keep_visible([Vector(-300, -100)], 10)
-        >>> print viewport.origin, viewport.scale
+        >>> print(viewport.origin, viewport.scale)
         (90.000, 190.000) 1.0
 
         >>> viewport.world_inner_bounds(10)
@@ -130,7 +131,7 @@ def doctest_Viewport_keep_visible():
         >>> viewport.keep_visible([Vector(-300, -100),
         ...                        Vector(500, 400)], 10)
 
-        >>> print viewport.origin, round(viewport.scale, 3)
+        >>> print(viewport.origin, round(viewport.scale, 3))
         (99.732, 190.000) 0.974
 
         >>> xmin, ymin, xmax, ymax = viewport.world_inner_bounds(10)
@@ -145,7 +146,7 @@ def doctest_Viewport_keep_visible():
 def doctest_FrameRateCounter_frame():
     """Tests for FrameRateCounter.frame
 
-        >>> from ui import FrameRateCounter
+        >>> from pyspacewar.ui import FrameRateCounter
         >>> frc = FrameRateCounter()
 
         >>> frc.frame()
@@ -155,7 +156,7 @@ def doctest_FrameRateCounter_frame():
         >>> len(frc.frames)
         2
 
-        >>> frc.frames = range(frc.avg_last_n_frames - 1)
+        >>> frc.frames = list(range(frc.avg_last_n_frames - 1))
         >>> frc.frame()
         >>> len(frc.frames) == frc.avg_last_n_frames
         True
@@ -175,7 +176,7 @@ def doctest_FrameRateCounter_frame():
 def doctest_FrameRateCounter_reset():
     """Tests for FrameRateCounter.reset
 
-        >>> from ui import FrameRateCounter
+        >>> from pyspacewar.ui import FrameRateCounter
         >>> frc = FrameRateCounter()
         >>> frc.frames = range(15)
         >>> frc.reset()
@@ -188,7 +189,7 @@ def doctest_FrameRateCounter_reset():
 def doctest_FrameRateCounter_fps():
     """Tests for FrameRateCounter.fps
 
-        >>> from ui import FrameRateCounter
+        >>> from pyspacewar.ui import FrameRateCounter
         >>> frc = FrameRateCounter()
         >>> frc.frames = []
         >>> frc.fps()
@@ -218,18 +219,20 @@ def doctest_FrameRateCounter_fps():
 
 def setUp(test=None):
     import pygame
-    pygame.init() # so that pygame.key.name() works
+    pygame.init()  # so that pygame.key.name() works
     # unfortunately, on linux, if $DISPLAY is unset, pygame.init doesn't
     # complain, but pygame.key.name() returns 'unknown key' for all keys
 
 
 def test_suite():
-    path = os.path.join(os.path.dirname(__file__), os.path.pardir)
+    path = os.path.normpath(
+        os.path.join(os.path.dirname(__file__), '..', '..'))
     if path not in sys.path:
         sys.path.append(path)
     return unittest.TestSuite([
-                        doctest.DocTestSuite('ui', setUp=setUp),
-                        doctest.DocTestSuite()])
+        doctest.DocTestSuite('pyspacewar.ui', setUp=setUp),
+        doctest.DocTestSuite(),
+    ])
 
 
 if __name__ == '__main__':